Gitit: update config to recognize where directories are
* static
* templates

we put these two directories under _wikidata_, update config
so that we don't have to use symbolic link.

Be ware of that, you should run gitit in the parent directory
of wikidata since gitit use the config value as a relative path
to where you are executing gitit.

for example, if you put wikidata to /home/foo/gitit/wikidata
and you are running gitit in /tmp/

gitit thinks static directory is in /tmp/wikidata/gitit/static